@@ -1,17 +1,17 @@
 'use strict';
 
-const vscode = require('vscode');
-const commandCheckoutAll = require('./command-checkoutall');
+import * as vscode from 'vscode';
+import commandCheckoutAll from './command-checkoutall';
 
 const COMMANDS = [commandCheckoutAll];
 
-exports.activate = function (context) {
+export function activate(context: vscode.ExtensionContext): void {
   for (const { commandId, execute } of COMMANDS) {
     const disposable = vscode.commands.registerCommand(commandId, execute);
     context.subscriptions.push(disposable);
   }
-};
+}
 
-exports.deactivate = function () {
+export function deactivate(): void {
   void 0;
-};
+}
